Skip spurious resilver IO on raidz vdev
On a raidz vdev, a block that does not span all child vdev's, excluding
its skip sectors if any, may not be affected by a child vdev outage or
failure. In such cases, the block does not need to be resilvered.
However, current resilver algorithm simply resilvers all blocks on a
degraded raidz vdev. Such spurious IO is not only wasteful, but also
adds the risk of overwriting good data.

This patch eliminates such spurious IOs.

Signed-off-by: Isaac Huang <he.huang@intel.com>